Moved transmission() and plane_parallel_rt() from spec_tools.py to radiative_transfer.py
